requirements.txt
# 核心深度学习框架
torch>=1.10.0

# 数据处理与科学计算
numpy
pandas
scipy
scikit-learn

# 配置文件解析
pyyaml

# 进度条显示
tqdm

# 绘图与可视化
matplotlib

# 其他工具
pathlib